eventlog: fixup eventlog tests and extend sample attester
The eventlog tests use a MockedAttester in place of a real one. After
introducing the CompositeEvidence struct, this becomes a little more
complicated. Instead, let's just use CompositeEvidence, which will
default to the sample attester.

To make sure that the sample attester will work with the tests, add the
runtime measurement methods to it.
